requirements.txt
requests
beautifulsoup4
pandas
matplotlib
numpy
python-docx
